refactor(app-shell): move sign-in panel and instance modal under account-popover
The account popover is the only consumer of both components; the
./instance-settings subpath export had zero importers. One directory now
owns the whole account surface, and one derivation owns the sync phase
presentation (the trigger tooltip and the popover status line previously
each switched on the phase).
