+++ packages/app-shell/src/account-popover/instance-settings-modal.svelte
