+++ packages/app-shell/src/account-popover/sign-in-panel.svelte
