--- packages/app-shell/src/instance-settings/index.ts
@@ -1,1 +1,0 @@
-export { default as InstanceSettingsModal } from './instance-settings-modal.svelte';
